requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
httpx
python-dotenv
pytest
pytest-asyncio
azure-ai-inference